Moving ACARS Hub logic out of application.py
